srpc-ctl: refactor srpc_ctl as srpc_controller
